App.vue split into components, api logic moved to pinia store.
