+++ src/main.ts
@@ -1,5 +1,9 @@
 import { createApp } from 'vue'
 import './style.css'
 import App from './App.vue'
+import { createPinia } from 'pinia'
+import piniaPluginPersistedState from 'pinia-plugin-persistedstate'
 
-createApp(App).mount('#app')
+createApp(App)
+.use(createPinia().use(piniaPluginPersistedState))
+.mount('#app')
